feat(dropout): 新增 elementwise_dropout 模块
- 实现 v_ref：使用 curandStateScrambledSobol32_t + curand_uniform 生成确定性随机数
- 标准 dropout 语义：概率 P 置零，存活元素缩放 1/(1-P)
- 全项目统一 get_FLOPs / get_bytes_transferred 返回类型为 long long，避免大尺寸溢出
- 项目链接 curand.lib，新增 CURAND_CHECK 错误检查宏
- 测试通过 CUB DeviceSelect 统计实际 zero 比率验证 dropout 概率
- 注册 type=10 入口，同步更新 kernel/test/main/test 声明

+++ src/copy_if.cu
@@ -5,12 +5,12 @@
 
 namespace copy_if 
 {
-    int get_FLOPs(const int size)
+    long long get_FLOPs(const long long size)
     {
         return size;
     }
 
-    int get_bytes_transferred(const int size)
+    long long get_bytes_transferred(const long long size)
     {
         return size * sizeof(float) + 2 * sizeof(int);
     }
