+++ src/quantize.cu
@@ -7,12 +7,12 @@
 
 namespace quantize
 {
-    int get_FLOPs(const int rows, const int cols)
+    long long get_FLOPs(const long long rows, const long long cols)
     {
         return rows * ((cols - 1) * 2 + 6 + 4 * cols); // 对于每一行，reduce max/min 均是 cols - 1 FLOPs，计算 scale 和 zeropoint 粗略认为是 6 FLOPs，最后 fp32 转 int8 粗略认为是 4 * cols FLOPs
     }
 
-    int get_bytes_transferred(const int rows, const int cols)
+    long long get_bytes_transferred(const long long rows, const long long cols)
     {
         return rows * cols * (sizeof(float) + sizeof(int8_t));
     }
