+++ src/reduce_sum.cu
@@ -5,12 +5,12 @@
 
 namespace reduce_sum 
 {
-    int get_FLOPs(const int size)
+    long long get_FLOPs(const long long size)
     {
         return size - 1;
     }
 
-    int get_bytes_transferred(const int size)
+    long long get_bytes_transferred(const long long size)
     {
         return (size + 1) * sizeof(float);
     }
